Implement custom filepicker, finish CLI acct controls
The filepicker native to bubbletea doesn't allow filtering within the
current folder, which makes it somewhat useless in directories with a
lot of files and folders. This adds a custom filepicker built on top of
the bubbletea List model, which does have support for filtering by a
struct's property (in this case, file or folder name).

A new endpoint has been added for fetching info about how the current
server has been configured, which can then be used for things like
determining if email signup is possible, checking if billing is set up,
etc.

CLI account controls have been finished up, allowing users to initiate
checkout from the CLI and not be required to log into the web interface.

+++ backend/server/payments/stripe/stripe.go
@@ -67,7 +67,7 @@ var stripeOrderTypeMap = map[string]string{
 
 // stripeProductStorage maps product IDs to their respective amounts of storage
 // that they grant a user
-var stripeProductStorage = map[string]int{
+var stripeProductStorage = map[string]int64{
 	config.YeetFileConfig.StripeBilling.SubNoviceMonthly: subscriptions.
 		StorageAmountMap[subscriptions.TypeNovice],
 	config.YeetFileConfig.StripeBilling.SubNoviceYearly: subscriptions.
@@ -86,7 +86,7 @@ var stripeProductStorage = map[string]int{
 
 // stripeProductSend maps product IDs to their respective amounts of storage
 // that they grant a user
-var stripeProductSend = map[string]int{
+var stripeProductSend = map[string]int64{
 	config.YeetFileConfig.StripeBilling.SubNoviceMonthly: subscriptions.
 		SendAmountMap[subscriptions.TypeNovice],
 	config.YeetFileConfig.StripeBilling.SubNoviceYearly: subscriptions.
